[FEATURE] Cockroachdb (#28)
* [FEATURE] New database

* [REFACTOR] Clear useless file

* [REFACTOR] Git repository test warnings

* [REFACTOR] Warning for unused fonction

* [DOC] Added Better Cluster repository documentation

* [BUGFIX] Schema.rs work with cockroachdb

* [FEATURE] Added production DockerFile

* [BUGFIX] Added cockroachdb in installation

+++ migrations/2022-06-17-122356_create_cargos/up.sql
@@ -5,7 +5,7 @@ create table "cargoes" (
   "name" VARCHAR NOT NULL,
   "image_name" VARCHAR NOT NULL,
   "binds" TEXT[] NOT NULL,
-  "replicas" INT NOT NULL DEFAULT 1 CHECK (replicas >= 0),
+  "replicas" BIGINT NOT NULL DEFAULT 1 CHECK (replicas >= 0),
   "dns_entry" VARCHAR,
   "domainname" VARCHAR,
   "hostname" VARCHAR
